fix preview of videos on add image page
yes im lazy
